feat: Add nix.settings.auto-optimise-store and remove unecessary comments.
nix.settings.auto-optimise-store:
If set to true, Nix automatically detects files in the store that have identical contents, and replaces them with hard links to a single copy. This saves disk space. If set to false (the default), you can still run nix-store –optimise to get rid of duplicate files.